feat: refactor codeMirror
